Try again

+++ src/main/java/com/example/shopapp/dtos/OrderDetailDTO.java
@@ -27,7 +27,7 @@ public class OrderDetailDTO {
 
     @JsonProperty("total_money")
     @Min(value = 0, message = "total_money must be >= 0")
-    private int totalMoney;
+    private Float totalMoney;
 
     private String color;
